requirements.txt
flask>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
deep-translator>=1.11.0
deepl>=1.17.0
apscheduler>=3.10.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
feedparser>=6.0.0
